% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_column_blobs.R
\name{get_column_blobs}
\alias{get_column_blobs}
\title{Find Connected Cells along Matrix Columns}
\usage{
get_column_blobs(m, offset = 0L)
}
\arguments{
\item{m}{a matrix of logical}

\item{offset}{integer to be added to the numbers indicating connected cells}
}
\value{
matrix of integer having the same dimension as \code{m}. Cells being
  \code{FALSE} in \code{m} are \code{0} in the output. Cells being
  \code{TRUE} in \code{m} are a positive integer number in the output.
  Connected cells within same columns share the same number.
}
\description{
Find Connected Cells along Matrix Columns
}
\examples{
column_blobs <- findblobs:::get_column_blobs(
  matrix(ncol = 3, byrow = TRUE, c(
    FALSE, TRUE, FALSE,
    TRUE,  TRUE, FALSE,
    FALSE, FALSE, TRUE,
    TRUE,  TRUE,  TRUE,
    TRUE,  FALSE, TRUE
  ))
)

findblobs::plot_integer_matrix(column_blobs)

}
