% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{order_by_first}
\alias{order_by_first}
\title{Order List of Vectors by First Vector Elements}
\usage{
order_by_first(x)
}
\arguments{
\item{x}{list of vectors}
}
\description{
Order List of Vectors by First Vector Elements
}
\examples{
findblobs:::order_by_first(list(
  c(3, 4),
  c(1, 5, 6),
  c(2, 1)
))

}
