% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_random_blobs.R
\name{place_random_blobs}
\alias{place_random_blobs}
\title{Create Random Blobs in a Matrix}
\usage{
place_random_blobs(m = matrix(0L, nrow = 10, ncol = 15), n_blobs = 5,
  min_fields = 3, max_fields = min_fields, do_plot = TRUE)
}
\arguments{
\item{m}{matrix of integer in which to place blobs. Default: 10 x 15-Matrix}

\item{n_blobs}{number of blobs to create. Default: 5}

\item{min_fields}{minimum number of fields a blob should consist of}

\item{max_fields}{maximum number of fields a blob should consist of. Defaults
to \code{min_fields}}

\item{do_plot}{if \code{TRUE} (the default), the result matrix is plotted}
}
\value{
matrix in which fields belonging to the same blob have the same
  integer number and zeros indicate empty fields
}
\description{
Create Random Blobs in a Matrix
}
\examples{
set.seed(42)
place_random_blobs(n_blobs = 5, min_fields = 5, max_fields = 10)

}
