% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_integer_matrix.R
\name{plot_integer_matrix}
\alias{plot_integer_matrix}
\title{Plot Matrix of Integer as Coloured Squares}
\usage{
plot_integer_matrix(x)
}
\arguments{
\item{x}{matrix of integer}
}
\description{
Plot Matrix of Integer as Coloured Squares
}
\examples{
plot_integer_matrix(matrix(nrow = 5, byrow = TRUE, c(
  2, 2, 2, 2, 2,
  2, 0, 1, 0, 2,
  2, 1, 1, 1, 2,
  2, 0, 1, 0, 2,
  2, 2, 2, 2, 2
)))

}
