% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{random_matrix}
\alias{random_matrix}
\title{Create Matrix with Randomly "Filled" Fields}
\usage{
random_matrix(matrix_dim = c(10, 10))
}
\arguments{
\item{matrix_dim}{numeric vector of length two giving the number of rows and
columns, respectively of the matrix}
}
\description{
Create Matrix with Randomly "Filled" Fields
}
